requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
redis==5.0.1
brawlstats==3.2.8